move CatalogDBObject-like classes to DatabaseEmulator
