+++ GCRCatalogs/gcrCatSimDatabaseEmulator.py
@@ -6,9 +6,10 @@ from collections import OrderedDict
 import numpy as np
 import gc
 import numbers
+from lsst.sims.utils import radiansFromArcsec
 
 
-__all__ = ["DESCQAObject"]
+__all__ = ["DESCQAObject", "bulgeDESCQAObject", "diskDESCQAObject"]
 
 
 _GCR_IS_AVAILABLE = True
@@ -26,6 +27,13 @@ except ImportError:
     from astropy.coordinates import SkyCoord
     def _angularSeparation(ra1, dec1, ra2, dec2):
         return SkyCoord(ra1, dec1, unit="deg").separation(SkyCoord(ra2, dec2, unit="deg")).radian
+
+def deg_to_radians(x):
+    return np.radians(x).astype(np.float)
+
+def arcsec_to_radians(x):
+    return radiansFromArcsec(x).astype(np.float)
+
 
 # a cache to store loaded catalogs to prevent them
 # from being loaded more than once, eating up
@@ -275,3 +283,56 @@ class DESCQAObject(object):
 
         return DESCQAChunkIterator(self._catalog, self.columnMap, obs_metadata,
                                    colnames, default, chunk_size)
+
+
+class bulgeDESCQAObject(DESCQAObject):
+    epoch = 2000.0
+
+    # PhoSim uniqueIds are generated by taking
+    # source catalog uniqueIds, multiplying by
+    # 1024, and adding objectTypeId.  This
+    # components of the same galaxy to have
+    # different uniqueIds, even though they
+    # share a uniqueId in the source catalog
+    objectTypeId = 77
+
+    # this identifies the uniqueId column
+    # in the source catalog
+    idColKey = 'galaxy_id'
+
+    # map the source catalog columns to CatSim
+    # columns; each tuple is :
+    #
+    # (catsim_column_name,
+    #  source_column_name,
+    #  optional transformation to get from one to the other)
+    columns = [('raJ2000', 'ra_true', deg_to_radians),
+               ('decJ2000', 'dec_true', deg_to_radians),
+               ('shear1', 'shear_1'),
+               ('shear2', 'shear_2'),
+               ('kappa', 'convergence'),
+               ('redshift', 'redshift_true'),
+               ('positionAngle', 'morphology/positionAngle'),
+               ('majorAxis', 'morphology/spheroidMajorAxisArcsec', arcsec_to_radians),
+               ('minorAxis', 'morphology/spheroidMinorAxisArcsec', arcsec_to_radians)]
+
+    # default values that are applied if no other source
+    # for a column can be found
+    dbDefaultValues = {'sindex': 4.0}
+
+
+class diskDESCQAObject(DESCQAObject):
+    epoch = 2000.0
+    objectTypeId = 87
+    idColKey = 'galaxy_id'
+    columns = [('raJ2000', 'ra_true', deg_to_radians),
+               ('decJ2000', 'dec_true', deg_to_radians),
+               ('shear1', 'shear_1'),
+               ('shear2', 'shear_2'),
+               ('kappa', 'convergence'),
+               ('redshift', 'redshift_true'),
+               ('positionAngle', 'morphology/positionAngle'),
+               ('majorAxis', 'morphology/diskMajorAxisArcsec', arcsec_to_radians),
+               ('minorAxis', 'morphology/diskMinorAxisArcsec', arcsec_to_radians)]
+
+    dbDefaultValues = {'sindex': 1.0}
