+++ examples/phosim_descqa.py
@@ -36,67 +36,8 @@ obs = obs_list[0]
 ############################################
 # establish connection to the DESCQA catalog
 
-from GCRCatalogs import DESCQAObject
+from GCRCatalogs import bulgeDESCQAObject, diskDESCQAObject
 from GCRCatalogs import sed_from_galacticus_mags
-from lsst.sims.utils import radiansFromArcsec
-
-def deg_to_radians(x):
-    return np.radians(x).astype(np.float)
-
-def arcsec_to_radians(x):
-    return radiansFromArcsec(x).astype(np.float)
-
-class bulgeDESCQAObject(DESCQAObject):
-    epoch = 2000.0
-
-    # PhoSim uniqueIds are generated by taking
-    # source catalog uniqueIds, multiplying by
-    # 1024, and adding objectTypeId.  This
-    # components of the same galaxy to have
-    # different uniqueIds, even though they
-    # share a uniqueId in the source catalog
-    objectTypeId = 77
-
-    # this identifies the uniqueId column
-    # in the source catalog
-    idColKey = 'galaxy_id'
-
-    # map the source catalog columns to CatSim
-    # columns; each tuple is :
-    #
-    # (catsim_column_name,
-    #  source_column_name,
-    #  optional transformation to get from one to the other)
-    columns = [('raJ2000', 'ra_true', deg_to_radians),
-               ('decJ2000', 'dec_true', deg_to_radians),
-               ('shear1', 'shear_1'),
-               ('shear2', 'shear_2'),
-               ('kappa', 'convergence'),
-               ('redshift', 'redshift_true'),
-               ('positionAngle', 'morphology/positionAngle'),
-               ('majorAxis', 'morphology/spheroidMajorAxisArcsec', arcsec_to_radians),
-               ('minorAxis', 'morphology/spheroidMinorAxisArcsec', arcsec_to_radians)]
-
-    # default values that are applied if no other source
-    # for a column can be found
-    dbDefaultValues = {'sindex': 4.0}
-
-class diskDESCQAObject(DESCQAObject):
-    epoch = 2000.0
-    objectTypeId = 87
-    idColKey = 'galaxy_id'
-    columns = [('raJ2000', 'ra_true', deg_to_radians),
-               ('decJ2000', 'dec_true', deg_to_radians),
-               ('shear1', 'shear_1'),
-               ('shear2', 'shear_2'),
-               ('kappa', 'convergence'),
-               ('redshift', 'redshift_true'),
-               ('positionAngle', 'morphology/positionAngle'),
-               ('majorAxis', 'morphology/diskMajorAxisArcsec', arcsec_to_radians),
-               ('minorAxis', 'morphology/diskMinorAxisArcsec', arcsec_to_radians)]
-
-    dbDefaultValues = {'sindex': 1.0}
-
 
 cat_file = 'proto-dc2_v2.0'
 db_bulge = bulgeDESCQAObject(cat_file)
